propeller/baseline-tfidf/requirements.txt
docx2txt>=0.8
langchain>=0.1.0
langchain-community>=0.0.11
networkx>=3.2.1
openpyxl>=3.1.2
pandas>=2.1.4
python-pptx>=0.6.23
scikit-learn>=1.3.2
unstructured>=0.12.0
xlrd>=2.0.1

pdfminer.six>=20231228
